[#28] Better handling of dir removal
When a directory is removed and we are in direct, don't blindly remove the collection.  Other objects may exist in the collection.  Instead, just remove the metadata with the fidstr of the deleted directory.

When adding a directory, if the directory already exists just continue to adding the metadata with the fidstr.